refactor(graphs): move types from models/ to root level
Simplify directory structure by moving chart-types.ts from models/
subdirectory to types.ts at the graphs root level. Update all imports
across graph components for better accessibility and organization.
